Created collector.go
